rename microapp
