+++ microapps/ArTabletopApp/app/proguard-rules.pro
