+++ microapps/ArTabletopApp/app/src/main/res/drawable/ic_launcher_background.xml
