+++ microapps/ArTabletopApp/app/src/main/res/values/colors.xml
