+++ settings.gradle.kts
@@ -139,5 +139,5 @@ include(":popup-app")
 project(":popup-app").projectDir = File(rootDir, "microapps/PopupApp/app")
 include(":ar")
 project(":ar").projectDir = File(rootDir, "toolkit/ar")
-include(":tabletop-ar-app")
-project(":tabletop-ar-app").projectDir = File(rootDir, "microapps/TabletopArApp/app")
+include(":ar-tabletop-app")
+project(":ar-tabletop-app").projectDir = File(rootDir, "microapps/ArTabletopApp/app")
